+++ src/modules/config.ts
@@ -61,7 +61,7 @@ function GetWorkspaceConfig(): UserConfig {
         remotePath: conf.get("remotePath", 'MES'),
         downloadOnOpen: false,
         uploadOnSave: false,
-        ignore: conf.get('ignore', ['*.json']),
+        ignore: conf.get('ignore', ['package.json']),
         useRootConfig: conf.get('useRootConfig', false),
         rootConfig: conf.get('rootConfig', '')
     };